feat(proving): bind verified ceremony evidence
