+++ cli/src/commands/trusted_setup.rs
@@ -15,7 +15,11 @@ pub fn package(
     phase1_blake2b512: &str,
     contributors: &[String],
     beacon_source: &str,
+    beacon_round: u64,
     beacon_randomness: &str,
+    beacon_evidence_sha256: &str,
+    beacon_commitment_publication: &str,
+    beacon_commitment_sha256: &str,
     beacon_iterations: u32,
     beacon_contribution_hash: &str,
     format: &str,
@@ -33,7 +37,11 @@ pub fn package(
         final_beacon: PackageFinalBeacon {
             contributed_zkey: Path::new(contributed_zkey),
             source: beacon_source,
+            round: beacon_round,
             randomness: beacon_randomness,
+            evidence_sha256: beacon_evidence_sha256,
+            commitment_publication: beacon_commitment_publication,
+            commitment_sha256: beacon_commitment_sha256,
             iterations: beacon_iterations,
             contribution_hash: beacon_contribution_hash,
         },
