+++ cli/src/main.rs
@@ -60,7 +60,11 @@ fn main() -> Result<()> {
                 phase1_blake2b512,
                 contributor,
                 beacon_source,
+                beacon_round,
                 beacon_randomness,
+                beacon_evidence_sha256,
+                beacon_commitment_publication,
+                beacon_commitment_sha256,
                 beacon_iterations,
                 beacon_contribution_hash,
                 format,
@@ -78,7 +82,11 @@ fn main() -> Result<()> {
             phase1_blake2b512,
             contributor,
             beacon_source,
+            *beacon_round,
             beacon_randomness,
+            beacon_evidence_sha256,
+            beacon_commitment_publication,
+            beacon_commitment_sha256,
             *beacon_iterations,
             beacon_contribution_hash,
             format,
