+++ proving/src/trusted_setup/support.rs
@@ -1,0 +1,58 @@
+use std::fs::File;
+use std::io::{Read, Seek, SeekFrom};
+use std::path::Path;
+
+use sha2::{Digest, Sha256};
+
+pub(super) fn open_regular_file(path: &Path, label: &str) -> Result<File, String> {
+    let metadata = std::fs::symlink_metadata(path)
+        .map_err(|error| format!("cannot inspect {label} `{}`: {error}", path.display()))?;
+    if metadata.file_type().is_symlink() || !metadata.is_file() {
+        return Err(format!(
+            "{label} `{}` must be a regular file",
+            path.display()
+        ));
+    }
+    File::open(path).map_err(|error| format!("cannot open {label} `{}`: {error}", path.display()))
+}
+
+pub(super) fn sha256_reader(reader: &mut File) -> Result<String, String> {
+    reader
+        .seek(SeekFrom::Start(0))
+        .map_err(|error| format!("cannot seek artifact for hashing: {error}"))?;
+    let mut hasher = Sha256::new();
+    let mut buffer = vec![0u8; 1024 * 1024];
+    loop {
+        let count = reader
+            .read(&mut buffer)
+            .map_err(|error| format!("cannot hash artifact: {error}"))?;
+        if count == 0 {
+            break;
+        }
+        hasher.update(&buffer[..count]);
+    }
+    Ok(hex_digest(hasher.finalize()))
+}
+
+pub(super) fn sha256_bytes(bytes: &[u8]) -> String {
+    hex_digest(Sha256::digest(bytes))
+}
+
+fn hex_digest(bytes: impl AsRef<[u8]>) -> String {
+    bytes
+        .as_ref()
+        .iter()
+        .map(|byte| format!("{byte:02x}"))
+        .collect()
+}
+
+pub(super) fn validate_hex(value: &str, length: usize, label: &str) -> Result<(), String> {
+    if value.len() != length
+        || !value
+            .bytes()
+            .all(|byte| byte.is_ascii_digit() || (b'a'..=b'f').contains(&byte))
+    {
+        return Err(format!("{label} must be {length} lowercase hex characters"));
+    }
+    Ok(())
+}
