+++ proving/tests/trusted_setup.rs
@@ -87,7 +87,7 @@ impl FixtureStore {
         let contribution_hash = "c4c8d61b26566a4e3110cb82dc894bdd5621c80d8d4e3d60b12671e6bb215413beebcb0b38cf77a2c77bb34736e1827ef1a65b9bc3694d6a6738867dead458c3";
         let transcript_value = serde_json::json!({
             "format": "achronyme-ceremony-transcript",
-            "version": 2,
+            "version": 3,
             "protocol": "groth16",
             "curve": "bn254",
             "circuit": {
@@ -114,7 +114,11 @@ impl FixtureStore {
             }],
             "final_beacon": {
                 "source": "https://example.invalid/public-randomness/42",
+                "round": 42,
                 "randomness": BEACON_RANDOMNESS,
+                "evidence_sha256": "eeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeee",
+                "commitment_publication": "https://example.invalid/test-only/commitment-42",
+                "commitment_sha256": "cccccccccccccccccccccccccccccccccccccccccccccccccccccccccccccccc",
                 "iterations": 10,
                 "contribution_hash": BEACON_CONTRIBUTION_HASH,
                 "contributed_zkey_sha256": CONTRIBUTED_ZKEY_SHA256
@@ -130,7 +134,7 @@ impl FixtureStore {
         std::fs::write(artifact_dir.join("transcript.json"), &transcript).unwrap();
         let manifest = serde_json::json!({
             "format": "achronyme-trusted-key",
-            "version": 2,
+            "version": 3,
             "protocol": "groth16",
             "curve": "bn254",
             "r1cs_sha256": digest,
@@ -148,7 +152,11 @@ impl FixtureStore {
                 }],
                 "final_beacon": {
                     "source": "https://example.invalid/public-randomness/42",
+                    "round": 42,
                     "randomness": BEACON_RANDOMNESS,
+                    "evidence_sha256": "eeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeee",
+                    "commitment_publication": "https://example.invalid/test-only/commitment-42",
+                    "commitment_sha256": "cccccccccccccccccccccccccccccccccccccccccccccccccccccccccccccccc",
                     "iterations": 10,
                     "contribution_hash": BEACON_CONTRIBUTION_HASH,
                     "contributed_zkey_sha256": CONTRIBUTED_ZKEY_SHA256
