+++ test/proving/documentation_contract.sh
@@ -27,10 +27,13 @@ rg -Fq 'Proving setup policy' "$repo_root/docs/migration/0.1.0.md"
 rg -Fq 'record-export-evidence.sh' "$trusted"
 rg -Fq -- '--export-evidence' "$trusted"
 rg -Fq -- '--contributed-zkey' "$trusted"
-rg -Fq -- '--beacon-source' "$trusted"
+rg -Fq 'commit-beacon.mjs' "$trusted"
+rg -Fq 'fetch-beacon.mjs' "$trusted"
+rg -Fq -- '--beacon-evidence' "$trusted"
+rg -Uiq 'publish.*commitment.*before.*target' "$trusted"
 rg -Uiq 'public beacon\s+does not replace\s+an\s+independent' "$trusted"
 rg -Fq -- '--contributed-zkey' "$boss"
-rg -Fq -- '--beacon-source' "$boss"
+rg -Fq -- '--beacon-evidence' "$boss"
 rg -Fq 'trusted-setup.md' "$repo_root/README.md"
 
 if rg -Fq 'Caching de `.zkey`' "$repo_root/STRATEGY.md"; then
