+++ test/proving/trusted_setup_workflow.sh
@@ -19,6 +19,8 @@ fi
 require_regular_file "$ach_bin" "Achronyme test binary"
 require_snarkjs
 require_command jq
+require_command node
+require_command npm
 
 work_root=$(mktemp -d)
 cleanup() {
@@ -44,11 +46,17 @@ evidence_checkout="$work_root/evidence-checkout"
 mkdir -p "$evidence_checkout/scripts/proving"
 cp "$scripts_dir/common.sh" "$scripts_dir/finalize-phase2.sh" \
     "$scripts_dir/record-export-evidence.sh" "$evidence_checkout/scripts/proving/"
+mkdir "$evidence_checkout/scripts/proving/drand"
+cp "$scripts_dir/drand"/*.mjs "$scripts_dir/drand"/package*.json \
+    "$evidence_checkout/scripts/proving/drand/"
+cp "$repo_root/.gitignore" "$evidence_checkout/.gitignore"
 git init -q "$evidence_checkout"
 git -C "$evidence_checkout" config user.email test@example.invalid
 git -C "$evidence_checkout" config user.name "Achronyme test"
-git -C "$evidence_checkout" add scripts
+git -C "$evidence_checkout" add .gitignore scripts
 git -C "$evidence_checkout" commit -qm initial
+npm ci --prefix "$evidence_checkout/scripts/proving/drand" --ignore-scripts \
+    >/dev/null
 recorded_export="$work_root/recorded-export.json"
 "$evidence_checkout/scripts/proving/record-export-evidence.sh" \
     --ach-bin "$ach_bin" --source "$source_file" --input-file "$input_file" \
@@ -94,8 +102,8 @@ snarkjs zkey verify "$r1cs" "$phase1" "$contributed_zkey" \
 contribution=$(extract_zkey_contributions "$work_root/test-zkey-verify.log" | sed -n '1p')
 [[ "$contribution" == *"|"* ]] || die "test contribution hash was not parsed"
 contributor="${contribution/|/=}"
-beacon_source="https://example.invalid/public-randomness/42"
-beacon_randomness=abcdef0123456789abcdef0123456789abcdef0123456789abcdef0123456789
+beacon_evidence="$repo_root/test/proving/fixtures/drand_quicknet_round_31006463.json"
+beacon_evidence_sha256=$(sha256_file "$beacon_evidence")
 
 export_evidence="$recorded_export"
 finalizer="$evidence_checkout/scripts/proving/finalize-phase2.sh"
@@ -103,8 +111,7 @@ finalizer="$evidence_checkout/scripts/proving/finalize-phase2.sh"
 declare -a finalize_args=(
     --r1cs "$r1cs" --wtns "$wtns" --phase1 "$phase1"
     --contributed-zkey "$contributed_zkey"
-    --beacon-source "$beacon_source" --beacon-randomness "$beacon_randomness"
-    --beacon-iterations 10
+    --beacon-evidence "$beacon_evidence" --beacon-iterations 10
     --source "$source_file" --input-file "$input_file"
     --ach-bin "$ach_bin" --phase1-source "https://example.invalid/test-only.ptau"
     --phase1-blake2b512 "$phase1_blake2b512" --contributor "$contributor"
@@ -127,15 +134,19 @@ grep -Fq 'export evidence witness SHA-256 mismatch' "$work_root/tampered.err"
     --export-evidence "$export_evidence" --work-dir "$ceremony_dir" --store "$store"
 
 evidence="$ceremony_dir/release-evidence.json"
-jq -e '
+jq -e --arg beacon_evidence_sha256 "$beacon_evidence_sha256" '
     .format == "achronyme-proving-release-evidence" and
-    .version == 2 and
+    .version == 3 and
     (.git_commit | test("^[0-9a-f]{40,64}$")) and
     (.achronyme_binary_sha256 | test("^[0-9a-f]{64}$")) and
     (.export_evidence.sha256 | test("^[0-9a-f]{64}$")) and
     .circuit.constraints == 1 and
-    .ceremony.final_beacon.source == "https://example.invalid/public-randomness/42" and
-    .ceremony.final_beacon.randomness == "abcdef0123456789abcdef0123456789abcdef0123456789abcdef0123456789" and
+    .ceremony.final_beacon.source == "https://api.drand.sh/52db9ba70e0cc0f6eaf7803dd07447a1f5477735fd3f661792ba94600c84e971/public/31006463" and
+    .ceremony.final_beacon.round == 31006463 and
+    .ceremony.final_beacon.randomness == "06664dcb57258c3ad1142e1f19575f3e597d29ee8eb49e957355dbab9d6935c9" and
+    .ceremony.final_beacon.evidence_sha256 == $beacon_evidence_sha256 and
+    .ceremony.final_beacon.commitment_publication == "https://example.invalid/test-only/commitment-31006463" and
+    .ceremony.final_beacon.commitment_sha256 == "54703c6c0df8236e97524ec5f0aaa8733566cf8131188d280a84b9b3f0e18a59" and
     .ceremony.final_beacon.iterations == 10 and
     (.ceremony.final_beacon.contribution_hash | test("^[0-9a-f]{128}$")) and
     (.metrics | length >= 12) and
@@ -146,8 +157,10 @@ require_regular_file "$store/$r1cs_sha256/manifest.json" "trusted-key manifest"
 require_regular_file "$store/$r1cs_sha256/transcript.json" "ceremony transcript"
 require_regular_file "$store/$r1cs_sha256/proving_key.zkey" "trusted proving key"
 jq -e '
-    .version == 2 and
-    .ceremony.final_beacon.source == "https://example.invalid/public-randomness/42"
+    .version == 3 and
+    .ceremony.final_beacon.round == 31006463 and
+    .ceremony.final_beacon.commitment_publication == "https://example.invalid/test-only/commitment-31006463" and
+    (.ceremony.final_beacon.evidence_sha256 | test("^[0-9a-f]{64}$"))
 ' "$store/$r1cs_sha256/manifest.json" >/dev/null
 
 printf 'trusted setup workflow test passed\n'
